refactor: fix PR #66 comments
